refactor(runtime): extract provider error classification
Move provider error evidence normalization, overflow patterns, classification, and presentation into a dependency-free leaf shared by ToolRuntime and ModelAdapter.

Relocate the taxonomy tests and add an extraction contract to keep the boundary explicit. Classification order and presentation strings are unchanged.
